Mejoras en la estructura de archivos y directorios, actualización del archivo README.md y corrección de tests
- Reorganización de la estructura de archivos y directorios.
- Actualización del archivo README.md.
- Corrección de errores en los tests unitarios.
