+++ README.md
@@ -1,6 +1,6 @@
 # My HTML to JSON Blocks Library
 
-**Descripción:**  
+**Descripción:**
 Esta librería convierte contenido HTML en una estructura de bloques JSON, útil para sistemas que requieren transformar contenido web en formatos procesables. Es altamente personalizable, permitiendo a los usuarios definir su propio método de conversión para imágenes mediante el uso de un sistema de clases abstractas.
 
 **Características principales:**
@@ -16,15 +16,26 @@ Esta librería convierte contenido HTML en una estructura de bloques JSON, útil
 - **Imágenes:** Proporciona un bloque de imagen con metadatos detallados, personalizable mediante el método abstracto `ImageTransformer`.
 
 **Cómo usar:**
-1. Instala las dependencias:
+1. Clona el repositorio:
+   ```bash
+   git clone https://github.com/yourusername/html-to-json-blocks.git
+   cd html-to-json-blocks
+   ```
+
+2. Instala las dependencias:
    ```bash
    pip install -r requirements.txt
    ```
 
-2. Crea una instancia del convertidor:
+3. Instala la librería:
+   ```bash
+   pip install .
+   ```
+
+4. Crea una instancia del convertidor:
    ```python
-   from my_html_to_json_lib.core import HtmlToJsonConverter
-   from my_html_to_json_lib.image_transformer import ImageTransformer
+   from html_to_json_blocks.core import HtmlToJsonConverter
+   from html_to_json_blocks.transformers.image_transformer import ImageTransformer
 
    class CustomImageTransformer(ImageTransformer):
        def transform_image(self, img_node, images_info):
@@ -34,10 +45,15 @@ Esta librería convierte contenido HTML en una estructura de bloques JSON, útil
    converter = HtmlToJsonConverter(image_transformer=CustomImageTransformer())
    ```
 
-3. Usa el convertidor para transformar HTML en bloques JSON:
+5. Usa el convertidor para transformar HTML en bloques JSON:
    ```python
+   from bs4 import BeautifulSoup
+
+   html_content = "<p>Ejemplo de contenido HTML</p>"
    soup = BeautifulSoup(html_content, 'html.parser')
+   images_info = []  # Información de imágenes si es necesario
    json_blocks = converter.convert(soup, images_info)
+   print(json_blocks)
    ```
 
 **Pruebas:**
@@ -49,4 +65,3 @@ pytest
 
 **Contribuciones:**
 Las contribuciones son bienvenidas. Por favor, abre un issue o envía un pull request si deseas mejorar la funcionalidad o corregir errores.
-
