+++ src/html_to_json_blocks/__init__.py
@@ -1,3 +1,5 @@
-# src/html_to_json_blocks/__init__.py
 from .core import HtmlToJsonConverter
 from .transformers.default_image_transformer import DefaultImageTransformer
+from .transformers.image_transformer import ImageTransformer
+
+__all__ = ["HtmlToJsonConverter", "DefaultImageTransformer", "ImageTransformer"]
