+++ tests/__init__.py
@@ -1,0 +1,5 @@
+# Making sure the unittest main runs the tests when the test suite is executed
+import unittest
+
+if __name__ == '__main__':
+    unittest.main()
